requirements.txt
Django==5.0.3
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4

